WBOITLayer cleanup
