+++ modules/layers/src/wboit-layer/wboit-layer-vertex-side.glsl.js
@@ -18,12 +18,12 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-import main from './solid-polygon-layer-vertex-main.glsl';
+import main from './wboit-layer-vertex-main.glsl';
 
 export default `\
 #version 300 es
 
-#define SHADER_NAME solid-polygon-layer-vertex-shader-side
+#define SHADER_NAME wboit-layer-vertex-shader-side
 #define IS_SIDE_VERTEX
